Massive bugfix and refactor
The internal logic is now much more consistent especially around
handling of what is a popwin and concretely defining it.

It also has proper handling of the case where a popterm/popwin is turned
into a regular split instead of a floating window.

It also handles well the case where the popwin is being occupied by
something that's not a popterm and closes it.